installers/tools/OSINT/sock-accounts-creator/requirements.txt
faker
pillow
requests
unidecode
pandas